requirements.txt
llama-index>=0.10.0
llama-index-llms-gemini
llama-index-embeddings-gemini
llama-index-vector-stores-qdrant
qdrant-client
ragas
streamlit
python-dotenv
pandas
nest_asyncio
langchain-google-genai
